Pull logic out to a claim button
