+++ src/components/Cards/AirdropCard/AirdropHeader.tsx
@@ -1,11 +1,7 @@
-import { getProof } from '@/async_fns/pandasia';
-import { useClaimAirdrop } from '@/async_fns/wagmi';
+import ClaimButton from '@/components/Button/ClaimButton';
 import Logo from '@/components/Logo';
-import { useC2PAuth } from '@/hooks/useC2PAuth';
 import { HexString } from '@/types/cryptoGenerics';
 import { format } from 'date-fns';
-import { useEffect, useState } from 'react';
-import { useContractWrite } from 'wagmi';
 
 type Props = {
   contractId: bigint;
@@ -24,28 +20,6 @@ export default function AirdropHeader({
   root,
   logo,
 }: Props) {
-  const { pChainAddr } = useC2PAuth();
-  const [proof, setProof] = useState<HexString[]>([]);
-
-  useEffect(() => {
-    fetchProof();
-  }, []);
-
-  async function fetchProof() {
-    const { data: proofy } = await getProof(root, pChainAddr || '', '');
-    if (proof === undefined) {
-      console.warn('proof undefined');
-      return;
-    }
-    setProof(proofy.Proof);
-  }
-
-  const { config, error, isError } = useClaimAirdrop(contractId, proof);
-  const { write: claim } = useContractWrite({ ...config });
-
-  const regex = /Error: (.*)/;
-  const errorMaybe = error?.message.match(regex);
-
   return (
     <div className="flex h-20 w-full items-center justify-between border-b border-secondary-700 p-6">
       <div className="flex items-center gap-2">
@@ -60,17 +34,7 @@ export default function AirdropHeader({
           </span>
         </div>
       </div>
-
-      <div>
-        {errorMaybe && <div>{errorMaybe[1]}</div>}
-        <button
-          onClick={claim}
-          className="border border-white px-4 py-2 text-xs font-semibold tracking-widest"
-          disabled={errorMaybe ? true : false}
-        >
-          CLAIM
-        </button>
-      </div>
+      <ClaimButton root={root} contractId={contractId} />
     </div>
   );
 }
